+++ R/densify.r
@@ -7,7 +7,7 @@
 #' points, i.e. new points will not fall on a great circle between existing
 #' vertices, rather they'll be along a straight line.
 #'
-#' @param x spatial features; lines or polygons from either the sf or sp
+#' @param x spatial features; lines or polygons from either the `sf` or `sp`
 #'   packages.
 #' @inheritParams smooth_densify
 #'