+++ NEWS.md
@@ -1,8 +1,10 @@
 # smoothr 0.0.1.9000
 
+- Drop small lines and polygons with `drop_crumbs()`
 - Kernel smoothing implemented in `smooth_ksmooth()`
 - Densification of spatial lines and polygons implemented with `densify()`
 - Raster example dataset added
+- Vignette updated to reflect new functionality
 
 # smoothr 0.0.1